Improve the performance of the Panic Goal mixin.
Store the block position via the block entity of the Will o' the Wisp
(this will need to be adapted to also work for the carved pine cone)
using `PathfinderMobData`.

This block position is then tested, if not null, against the entity
itself and, if it is further than 2 blocks away, set to null.

This moves the logic from each instance of the entity into the Will 'o
the Wisp block entity, dramatically improving performance.